# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Council Asset & Compliance Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/upload:
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document or evidence file
      parameters:
        - name: x-filename
          in: header
          schema: { type: string }
        - name: assetId
          in: query
          schema: { type: string }
        - name: complianceItemId
          in: query
          schema: { type: string }
        - name: jobId
          in: query
          schema: { type: string }
        - name: serviceRecordId
          in: query
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents by context (assetId, jobId, etc.)
      parameters:
        - name: assetId
          in: query
          schema: { type: string }
        - name: complianceItemId
          in: query
          schema: { type: string }
        - name: jobId
          in: query
          schema: { type: string }
        - name: serviceRecordId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /documents/{docId}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document metadata
      parameters:
        - name: docId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Soft-delete a document
      parameters:
        - name: docId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
  /documents/{docId}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Download or view a document file
      parameters:
        - name: docId
          in: path
          required: true
          schema: { type: string }
        - name: inline
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

# ── COMPONENTS ────────────────────────────────────────────────────────────
components:
  schemas:
...
    Document:
      type: object
      required: [id, tenantId, fileName, filePath, fileType, uploadedBy, createdAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        assetId: { type: ["string", "null"] }
        complianceItemId: { type: ["string", "null"] }
        jobId: { type: ["string", "null"] }
        serviceRecordId: { type: ["string", "null"] }
        fileName: { type: string }
        filePath: { type: string }
        fileType: { type: string }
        fileSize: { type: ["integer", "null"] }
        uploadedBy: { type: string }
        createdAt: { type: string }
        deletedAt: { type: ["string", "null"] }